requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
keras==2.12.0
language_tool_python==2.7.1
lexicalrichness==0.5.0
matplotlib==3.7.0
numpy==1.23.5
pandas==2.1.2
protobuf==3.20.3
PyPDF2==3.0.1
pyspellchecker==0.7.2
python_docx==1.0.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.2.2
seaborn==0.13.0
sentence_transformers==2.2.2
shap==0.43.0
spacy==3.5.2
tensorflow==2.12.0
tensorflow_intel==2.12.0
textblob==0.17.1
textstat==0.7.3
torch==2.0.0+cu117
tqdm==4.66.1
xgboost==1.7.5
